Add `human` format to test-tool
Add the human format support to the test tool so that TEST_DATE_NOW
can be used to specify the current time.

The get_time() helper function was created and and checks the
TEST_DATE_NOW environment variable.  If TEST_DATE_NOW is set, then
that date is used instead of the date returned by by gettimeofday().

All calls to gettimeofday() were replaced by calls to get_time().

Signed-off-by: Stephen P. Smith <ischis2@cox.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>